ASN-8 Full responsive landing page added, Button.css and Button.jsx moved to /components/ui, few changes in index.css and app.css
