+++ frontend/src/LandingPage.css
@@ -1,0 +1,67 @@
+body {
+  background-color: #dcf2fb;
+}
+
+.landing-container h1 {
+  margin-top: -50px;
+  color: black;
+  font-size: 60px;
+  font-family: "Koulen", sans-serif;
+  text-align: center;
+  display: flex;
+}
+
+.landing-buttons-container {
+  background-color: white;
+  height: 114px;
+  border-bottom: 3px solid #04387c;
+  position: fixed;
+  top: 0;
+  left: 0px;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  z-index: 1;
+  padding: 20px;
+}
+
+.landing-buttons-container > :last-child {
+  margin-right: 42px;
+}
+
+.landing-container {
+  background-color: white;
+  margin-top: 37px;
+  border-radius: 10px;
+  border: 4px solid #04387c;
+  width: 886px;
+  height: 582px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+}
+
+.landing-container p {
+  color: black;
+  font-size: 34px;
+  margin-left: 95px;
+  margin-right: 95px;
+  font-family: "Kumbh Sans", sans-serif;
+
+  text-align: center;
+}
+
+.landing-container .photo1 {
+  position: absolute;
+  top: 110px;
+  right: 11.5px;
+}
+
+.landing-container .photo2 {
+  position: absolute;
+  left: 20px;
+  bottom: 34.64px;
+}
