+++ frontend/src/components/ui/Button.css
@@ -6,10 +6,8 @@
   padding: 36px 249px;
   gap: 10px;
 
-  width: 498px;
-  height: 72px;
-  left: 20px;
-  top: 20px;
+  width: 22px;
+  height: 15px;
 
   border-radius: 14px;
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
@@ -23,18 +21,16 @@
   flex: none;
   order: 0;
   flex-grow: 0;
-
 }
-  
-  .blue {
-    background-color: #04387C;
-    color: #ffff;
-    white-space: nowrap;
-  }
-  
-  .purple {
-    background-color: #8B5BC8;
-    color: #ffffff;
-    white-space: nowrap;
-  }
-  
+
+.blue {
+  background-color: #04387c;
+  color: #ffff;
+  white-space: nowrap;
+}
+
+.purple {
+  background-color: #8b5bc8;
+  color: #ffffff;
+  white-space: nowrap;
+}
