+++ frontend/src/index.css
@@ -1,3 +1,5 @@
+@import url("https://fonts.googleapis.com/css2?family=Koulen&family=Kumbh+Sans&display=swap");
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
@@ -23,6 +25,7 @@ a:hover {
 }
 
 body {
+  font-family: "Kumbh Sans", sans-serif;
   margin: 0;
   display: flex;
   place-items: center;
